flatten the structure, remove dotfiles
dotfiles now live in a separate repo. This repo is meant to be cloned into
~bin directory, which is meant to be included in PATH environment variable.

TODO: set proper permissions in a post-checkout hook (or something; not sure,
where exactly) to mitigate a security hole.  The bin dir should not be writable
by others, the scripts in it (including sourced files) should not be writable
by anyone.